+++ src/datasources/products/ProductsDataSource.test.ts
@@ -740,6 +740,22 @@ describe('query', () => {
         { text: 'B Product (123)', value: '123' }
       ]);
     });
+
+    test('should run query if not hidden', () => {
+      const query: ProductQuery = {
+        hide: false,
+        refId: ''
+      };
+      expect(datastore.shouldRunQuery(query)).toBe(true);
+    });
+
+    test('should not run query if hidden', () => {
+      const query: ProductQuery = {
+        hide: true,
+        refId: ''
+      };
+      expect(datastore.shouldRunQuery(query)).toBe(false);
+    });
   });
 });