+++ src/datasources/products/ProductsDataSource.ts
@@ -165,7 +165,7 @@ export class ProductsDataSource extends DataSourceBase<ProductQuery> {
   }
 
   shouldRunQuery(query: ProductQuery): boolean {
-    return true;
+    return !query.hide;
   }
 
   async testDatasource(): Promise<TestDataSourceResponse> {